% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-vclSparseMatrix.R
\docType{class}
\name{dvclSparseMatrix-class}
\alias{dvclSparseMatrix-class}
\title{dvclSparseMatrix Class}
\description{
An integer type matrix in the S4 \code{vclSparseMatrix}
representation.
}
\section{Slots}{

 \describe{
     \item{\code{address}:}{Pointer to a double type matrix}
 }
}

\seealso{
\code{\link{vclSparseMatrix-class}},
\code{\link{fvclSparseMatrix-class}}
}
\author{
Charles Determan Jr.
}
