% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-vclSparseMatrix.R
\docType{class}
\name{fvclSparseMatrix-class}
\alias{fvclSparseMatrix-class}
\title{fvclSparseMatrix Class}
\description{
An integer type matrix in the S4 \code{vclSparseMatrix}
representation.
}
\section{Slots}{

 \describe{
     \item{\code{address}:}{Pointer to a float matrix.}
 }
}

\seealso{
\code{\link{vclSparseMatrix-class}},
\code{\link{dvclSparseMatrix-class}}
}
\author{
Charles Determan Jr.
}
