% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/typeof.R
\docType{methods}
\name{typeof,vclSparseMatrix-method}
\alias{typeof,vclSparseMatrix-method}
\title{Get gpuRsparse object type}
\usage{
\S4method{typeof}{vclSparseMatrix}(x)
}
\arguments{
\item{x}{A gpuRsparse object}
}
\description{
\code{typeof} determines the type (i.e. storage mode) of a
gpuRsparse object
}
\author{
Charles Determan Jr.
}
