% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-vclSparseMatrix.R
\docType{class}
\name{vclSparseMatrix-class}
\alias{vclSparseMatrix-class}
\title{vclSparseMatrix Class}
\description{
This is the 'mother' class for all
vclSparseMatrix objects.  These objects are pointers
to viennacl compressed matrices directly on the GPU.  This will
avoid the overhead of passing data back and forth
between the host and device.

As such, any changes made
to normal R 'copies' (e.g. A <- B) will be propogated to
the parent object.

There are multiple child classes that correspond
to the particular data type contained.  These include
\code{fvclSparseMatrix} and
\code{dvclSparseMatrix} corresponding to integer, float, and
double data types respectively.
}
\note{
R does not contain a native float type.  As such,
the matrix data within a \code{\link{fvclSparseMatrix-class}}
will be represented as double but downcast when any
vclSparseMatrix methods are used.

May also remove the type slot
}
\section{Slots}{

 Common to all vclSparseMatrix objects in the package
 \describe{
     \item{\code{address}:}{Pointer to data matrix}
     \item{\code{.context_index}:}{Integer index of OpenCL contexts}
     \item{\code{.platform_index}:}{Integer index of OpenCL platforms}
     \item{\code{.platform}:}{Name of OpenCL platform}
     \item{\code{.device_index}:}{Integer index of active device}
     \item{\code{.device}:}{Name of active device}
 }
}

\seealso{
\code{\link{fvclSparseMatrix-class}},
\code{\link{dvclSparseMatrix-class}}
}
\author{
Charles Determan Jr.
}
