% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vclSparseMatrix.R
\docType{methods}
\name{vclSparseMatrix}
\alias{vclSparseMatrix}
\alias{vclSparseMatrix,matrix-method}
\alias{vclSparseMatrix,matrix}
\title{Construct a vclSparseMatrix}
\usage{
vclSparseMatrix(data = NA, nrow = NA, ncol = NA, type = NULL, ...)

\S4method{vclSparseMatrix}{matrix}(data, type = NULL, ctx_id = NULL)
}
\arguments{
\item{data}{An object that is or can be converted to a
\code{matrix} object}

\item{nrow}{An integer specifying the number of rows}

\item{ncol}{An integer specifying the number of columns}

\item{type}{A character string specifying the type of vclSparseMatrix.  Default
is NULL where type is inherited from the source data type.}

\item{...}{Additional method to pass to vclSparseMatrix methods}

\item{ctx_id}{An integer specifying the object's context}
}
\value{
A vclSparseMatrix object
}
\description{
Construct a vclSparseMatrix of a class that inherits
from \code{vclSparseMatrix}.  This class points to memory directly on
the GPU to avoid the cost of data transfer between host and device.
}
\author{
Charles Determan Jr.
}
